Split source files into smaller parts to reduce redundancy
