+++ src/isolines.js
@@ -5,38 +5,28 @@
 * https://github.com/RaumZeit/MarchingSquares.js
 */
 
+import { optIsoLines } from './options.js';
+import { extractPolygons, traceLinePaths } from './polygons.js';
+
 
 /*
-  Compute the isocontour(s) of a scalar 2D field given
+  Compute the iso lines for a scalar 2D field given
   a certain threshold by applying the Marching Squares
   Algorithm. The function returns a list of path coordinates
 */
-var defaultSettings = {
-  successCallback:  null,
-  verbose:          false,
-  polygons:         false,
-  polygons_full:    true,
-  linearRing:       true,
-  interpolate:    function(a, b, threshold) {
-    if (a < b)
-      return (threshold - a) / (b - a);
-    else
-      return (a - threshold) / (a - b);
-  },
-};
-
-var settings = {};
 
 function isoLines(data, threshold, options){
+  var settings = {};
+
   /* process options */
   options = options ? options : {};
 
-  var optionKeys = Object.keys(defaultSettings);
+  var optionKeys = Object.keys(optIsoLines);
 
   for(var i = 0; i < optionKeys.length; i++){
     var key = optionKeys[i];
     var val = options[key];
-    val = ((typeof val !== 'undefined') && (val !== null)) ? val : defaultSettings[key];
+    val = ((typeof val !== 'undefined') && (val !== null)) ? val : optIsoLines[key];
 
     settings[key] = val;
   }
@@ -47,17 +37,33 @@ function isoLines(data, threshold, options){
   /* restore compatibility */
   settings.polygons_full  = !settings.polygons;
 
-  var grid = computeContourGrid(data, threshold, settings);
+  var grid = {
+    rows:       data.length - 1,
+    cols:       data[0].length - 1,
+    cells:      [],
+    threshold:  threshold
+  };
+
+  settings.threshold = threshold;
+
+  for (var j = 0; j < grid.rows; ++j) {
+    grid.cells[j] = [];
+    for (var i = 0; i < grid.cols; ++i)
+      grid.cells[j][i] = prepareCell(data, i, j, settings);
+  }
 
   var ret;
+
   if(settings.polygons){
     if (settings.verbose)
       console.log("MarchingSquaresJS-isoContours: returning single polygons for each grid cell");
-    ret = GetPolygons(grid, settings);
+
+    ret = extractPolygons(grid, settings);
   } else {
     if (settings.verbose)
       console.log("MarchingSquaresJS-isoContours: returning iso lines (polygon paths) for entire data grid");
-    ret = TracePaths(grid, settings);
+
+    ret = traceLinePaths(grid, settings);
   }
 
   if(typeof settings.successCallback === 'function')
@@ -510,298 +516,9 @@ function prepareCell(grid, x, y, settings) {
       }
 
       break;
-
   }
 
   return cell;
-}
-
-
-/* compute the isocontour 4-bit grid */
-function computeContourGrid(data, threshold, settings) {
-  var rows = data.length - 1;
-  var cols = data[0].length - 1;
-  var ContourGrid = { rows: rows, cols: cols, cells: [], threshold: threshold };
-
-  settings.threshold = threshold;
-
-  for (var j = 0; j < rows; ++j) {
-    ContourGrid.cells[j] = [];
-    for (var i = 0; i < cols; ++i)
-      ContourGrid.cells[j][i] = prepareCell(data, i, j, settings);
-  }
-
-  return ContourGrid;
-}
-
-function entry_coordinate(x, y, mode, path) {
-  var k = x;
-  var l = y;
-
-  if (mode === 0) { /* down */
-    k += 1;
-    l += path[0][1];
-  } else if (mode === 1) { /* left */
-    k += path[0][0];
-  } else if (mode === 2) { /* up */
-    l += path[0][1];
-  } else if (mode === 3) { /* right */
-    k += path[0][0];
-    l += 1;
-  }
-
-  return [ k, l ];
-}
-
-function skip_coordinate(x, y, mode) {
-  var k = x;
-  var l = y;
-
-  if (mode === 0) { /* down */
-    k++;
-  } else if (mode === 1) { /* left */
-    /* do nothing */
-  } else if (mode === 2) { /* up */
-    l++;
-  } else if (mode === 3) { /* right */
-    k++;
-    l++;
-  }
-
-  return [ k, l ];
-}
-
-function TracePaths(grid, settings) {
-  var areas = [];
-  var rows = grid.rows;
-  var cols = grid.cols;
-  var polygons = [];
-
-  /*
-      directions for out-of-grid moves are:
-      0 ... "down",
-      1 ... "left",
-      2 ... "up",
-      3 ... "right"
-  */
-  var valid_entries = [ "right",  /* down */
-                        "bottom", /* left */
-                        "left",   /* up */
-                        "top"     /* right */
-                      ];
-  var add_x         = [ 0, -1, 0, 1 ];
-  var add_y         = [ -1, 0, 1, 0 ];
-  var entry_dir     =  { bottom: 1,
-                         left: 2,
-                         top: 3,
-                         right: 0
-                       };
-
-  /* first, detect whether we need any outer frame */
-  var require_frame = true;
-
-  for (var j = 0; j < rows; j++) {
-    if ((grid.cells[j][0].x0 >= grid.threshold) ||
-        (grid.cells[j][cols - 1].x1 >= grid.threshold)) {
-      require_frame = false;
-      break;
-    }
-  }
-
-  if ((require_frame) &&
-      ((grid.cells[rows - 1][0].x3 >=  grid.threshold) ||
-      (grid.cells[rows - 1][cols - 1].x2 >= grid.threshold))) {
-    require_frame = false;
-  }
-
-  if (require_frame)
-    for (var i = 0; i < cols - 1; i++) {
-      if ((grid.cells[0][i].x1 >= grid.threshold) ||
-          (grid.cells[rows - 1][i].x2 > grid.threshold)) {
-        require_frame = false;
-        break;
-      }
-    }
-
-  if (require_frame) {
-    if (settings.linearRing)
-      polygons.push([ [0, 0], [0, rows], [cols, rows], [cols, 0], [0, 0] ]);
-    else
-      polygons.push([ [0, 0], [0, rows], [cols, rows], [cols, 0] ]);
-  }
-
-  /* finally, start tracing back first polygon(s) */
-  for (var j = 0; j < rows; j++) {
-    for (var i = 0; i < cols; i++) {
-      if (typeof grid.cells[j][i] !== 'undefined') {
-        var cell = grid.cells[j][i];
-        if (cell.cval === 15)
-          continue;
-
-        var nextedge = null;
-
-        /* trace paths for all available edges that go through this cell */
-        for (var e = 0; e < 4; e++) {
-          nextedge = valid_entries[e];
-
-          if (typeof cell.edges[nextedge] === 'object') {
-            /* start a new, full path */
-            var path              = [];
-            var ee                = cell.edges[nextedge];
-            var enter             = nextedge;
-            var x                 = i;
-            var y                 = j;
-            var finalized         = false;
-            var origin            = [ i + ee.path[0][0], j + ee.path[0][1] ];
-
-            /* add start coordinate */
-            path.push(origin);
-
-            /* start traceback */
-            while (!finalized) {
-              var cc = grid.cells[y][x];
-
-              if (typeof cc.edges[enter] !== 'object')
-                break;
-
-              var ee = cc.edges[enter];
-
-              /* remove edge from cell */
-              delete cc.edges[enter];
-
-              /* add last point of edge to path arra, since we extend a polygon */
-              var point = ee.path[1];
-              point[0] += x;
-              point[1] += y;
-              path.push(point);
-
-              enter = ee.move.enter;
-              x     = x + ee.move.x;
-              y     = y + ee.move.y;
-
-              /* handle out-of-grid moves */
-              if ((typeof grid.cells[y] === 'undefined') ||
-                  (typeof grid.cells[y][x] === 'undefined')) {
-
-                if (!settings.linearRing)
-                  break;
-
-                var dir   = 0;
-                var count = 0;
-
-                if (x === cols) {
-                  x--;
-                  dir = 0;  /* move downwards */
-                } else if (x < 0) {
-                  x++;
-                  dir = 2;  /* move upwards */
-                } else if (y === rows) {
-                  y--;
-                  dir = 3;  /* move right */
-                } else if (y < 0) {
-                  y++;
-                  dir = 1;  /* move left */
-                }
-
-                if ((x === i) && (y === j) && (dir === entry_dir[nextedge])) {
-                  finalized = true;
-                  enter     = nextedge;
-                  break;
-                }
-
-                while (1) {
-                  var found_entry = false;
-
-                  if (count > 4) {
-                    console.log("Direction change counter overflow! This should never happen!");
-                    break;
-                  }
-
-                  cc = grid.cells[y][x];
-
-                  /* check for re-entry */
-                  var ve = valid_entries[dir];
-                  if (typeof cc.edges[ve] === 'object') {
-                    /* found re-entry */
-                    ee = cc.edges[ve];
-                    path.push(entry_coordinate(x, y, dir, ee.path));
-                    enter = ve;
-                    found_entry = true;
-                    break;
-                  }
-
-                  if (found_entry) {
-                    break;
-                  } else {
-                    path.push(skip_coordinate(x, y, dir));
-
-                    x += add_x[dir];
-                    y += add_y[dir];
-
-                    /* change direction if we'e moved out of grid again */
-                    if ((typeof grid.cells[y] === 'undefined') || (typeof grid.cells[y][x] === 'undefined')) {
-                      x -= add_x[dir];
-                      y -= add_y[dir];
-
-                      dir = (dir + 1) % 4;
-                      count++;
-                    }
-
-                    if ((x === i) && (y === j) && (dir === entry_dir[nextedge])) {
-                      /* we are back where we started off, so finalize the polygon */
-                      finalized = true;
-                      enter     = nextedge;
-                      break;
-                    }
-                  }
-                }
-              }
-            }
-
-            if ((settings.linearRing) &&
-                ((path[path.length - 1][0] !== origin[0]) ||
-                (path[path.length - 1][1] !== origin[1])))
-              path.push(origin);
-
-            polygons.push(path);
-          }
-        }
-      }
-    }
-  }
-
-  return polygons;
-}
-
-function GetPolygons(grid, settings) {
-  var areas = [];
-  var rows = grid.rows;
-  var cols = grid.cols;
-  var polygons = [];
-
-  for (var j = 0; j < rows; j++) {
-    for (var i = 0; i < cols; i++) {
-      if (typeof grid.cells[j][i] !== 'undefined') {
-        var cell = grid.cells[j][i];
-        if (cell.cval === 15)
-          continue;
-
-        cell.polygons.forEach(function(p) {
-          p.forEach(function(pp) {
-            pp[0] += i;
-            pp[1] += j;
-          });
-
-          if (settings.linearRing)
-            p.push(p[0]);
-
-          polygons.push(p);
-        });
-      }
-    }
-  }
-
-  return polygons;
 }
